Doing some updates to the project's description
Based on what I remember about this project from a semester ago. I'll need to actually make a real update to this at some point that way I can have more confidence in showcasing this project. While I certainly learned a lot in the process of developing it, it's still not really in a state where it can be used.